Work with file paths, not URI substrings, in custom URI scheme
Using `…_get_uri()` gives a string that includes the anchor, which
isn't a valid entry path in the archive. `…_get_path()` gets just
the path (but with a leading slash).

Rather than special case the widget code to handle the starting slash,
the archive and doc now ignore the leading slash in paths. This allows
`…_get_path()` to be used anywhere in the code base without additional
text manipulation each time.